--- packages/server/src/app/api/search/models/index.ts
@@ -1,3 +1,0 @@
-export { Pagination, DEFAULT_PAGE_SIZE, MAX_PAGE_SIZE } from './pagination';
-export { SearchParameters } from './search-parameters';
-export { SearchResults } from './search-results';
