--- packages/server/src/app/api/search/models/search-parameters.ts
@@ -1,6 +1,0 @@
-import { Pagination } from './pagination';
-
-export interface SearchParameters {
-    text: string;
-    pagination: Pagination;
-}
