--- packages/server/src/app/api/search/models/search-results.ts
@@ -1,8 +1,0 @@
-import { Document } from 'mongoose';
-import { Pagination } from './pagination';
-
-export interface SearchResults<T extends Document> {
-    matches: T[];
-    pagination: Pagination;
-    totalPages: number;
-}
