--- packages/server/src/app/api/search/search.controller.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { SearchController } from './search.controller';
-
-describe('Search Controller', () => {
-    let controller: SearchController;
-
-    beforeEach(async () => {
-        const module: TestingModule = await Test.createTestingModule({
-            controllers: [SearchController],
-        }).compile();
-
-        controller = module.get<SearchController>(SearchController);
-    });
-
-    it('should be defined', () => {
-        expect(controller).toBeDefined();
-    });
-});
