--- packages/server/src/app/api/search/search.service.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { SearchService } from './search.service';
-
-describe('SearchService', () => {
-    let service: SearchService;
-
-    beforeEach(async () => {
-        const module: TestingModule = await Test.createTestingModule({
-            providers: [SearchService],
-        }).compile();
-
-        service = module.get<SearchService>(SearchService);
-    });
-
-    it('should be defined', () => {
-        expect(service).toBeDefined();
-    });
-});
