+++ packages/server/src/app/services/search/index.ts
@@ -1,0 +1,3 @@
+import { SearchService } from './search.service';
+
+export const SearchServices = [SearchService];
