+++ packages/server/src/app/services/search/search.service.ts
@@ -1,18 +1,18 @@
 import { Injectable } from '@nestjs/common';
 import { PaginateResult } from 'mongoose';
-
 import * as sanitizeHtml from 'sanitize-html';
+
+import { ISearch } from '../../shared/search';
 import { ContentFilter } from '@dragonfish/models/works';
 import { InitialResults } from '@dragonfish/models/util';
-
-import { UsersService } from '../../db/users/users.service';
-import { BlogsService, ContentService } from '../../db/content';
+import { UsersStore } from '../../db/users/users.store';
+import { ContentStore } from '../../db/content';
 import { User } from '@dragonfish/models/users';
-import { BlogsContentModel, ContentKind, ContentModel } from '@dragonfish/models/content';
+import { ContentKind, ContentModel } from '@dragonfish/models/content';
 
 @Injectable()
-export class SearchService {
-    constructor(private readonly usersService: UsersService, private readonly contentService: ContentService) {}
+export class SearchService implements ISearch {
+    constructor(private readonly usersService: UsersStore, private readonly contentService: ContentStore) {}
 
     async fetchInitialResults(query: string, contentFilter: ContentFilter): Promise<InitialResults> {
         const parsedQuery = `"${sanitizeHtml(query)}"`;
@@ -57,7 +57,7 @@ export class SearchService {
         );
     }
 
-    async searchWorks(
+    async searchContent(
         query: string,
         pageNum: number,
         contentFilter: ContentFilter,
