+++ packages/server/src/app/shared/search/index.ts
@@ -1,0 +1,1 @@
+export { ISearch } from './search.interface';
